backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pymongo==4.6.1
selenium==4.18.1
requests==2.31.0
beautifulsoup4==4.12.3
pytest==8.0.2
pytest-selenium==4.0.0
python-dotenv==1.0.1
gunicorn==21.2.0 